requirements.txt
# Flask Web Application Requirements
# Exact same packages as desktop version plus Flask web framework

# Flask web framework
Flask==3.0.0
Werkzeug==3.0.1

# Selenium web automation (same as desktop version)
selenium==4.15.2
webdriver-manager==4.0.1

# Image processing (same as desktop version)
Pillow==10.1.0

# Excel generation (same as desktop version)
openpyxl==3.1.2

# Additional Python packages for web deployment
gunicorn==21.2.0
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
pytest-flask==1.3.0
black==23.11.0
flake8==6.1.0

# Security and monitoring
cryptography==41.0.8